refactor: 에디터 페이지 사진 추가 컴포넌트 분리
